调整mediaplayer代码结构，根据功能划分为几个文件，编译并在模拟器运行通过
