+++ navigate/logicmacro.c
@@ -1,0 +1,25 @@
+#include "logicmacro.h"
+
+/*===========================================================================
+   This function releases IBase.
+===========================================================================*/
+void TS_FreeIF(IBase ** ppif)
+{
+   if (ppif && *ppif)
+   {
+      IBASE_Release(*ppif);
+      *ppif = NULL;
+   }
+}
+
+/*===========================================================================
+   This function releases IWindow.
+===========================================================================*/
+void TS_FreeWin(IWindow ** ppif)
+{
+   if (ppif && *ppif)
+   {
+      IWINDOW_Delete(*ppif);
+      *ppif = NULL;
+   }
+}
