+++ navigate/mainwindow.h
@@ -1,0 +1,10 @@
+#ifndef  TS_MAINWINDOW_H
+#define  TS_MAINWINDOW_H
+
+#include "navigatepch.h"
+
+typedef struct CTopSoupApp CTopSoupApp;
+
+// Main window
+IWindow *  CMainWin_New(CTopSoupApp * pOwner);
+#endif
