+++ navigate/navigatepch.h
@@ -1,0 +1,24 @@
+#ifndef TS_NAVIGATEPCH_H
+#define TS_NAVIGATEPCH_H
+
+#include "AEEAppGen.h"	      // AEEApplet structure and Applet services
+#include "AEEStdLib.h"        // AEE Stb Lib Services
+#include "AEEImageCtl.h"      // AEE ImageCtl Services
+#include "nmdef.h"
+#include "AEEMenu.h"  
+
+#include "window.h"
+#include "navigate.h"
+#include "mainwindow.h"
+#include "wincontrol.h"
+
+#include "navigate.bid"
+#include "navigate.brh"
+
+	
+#include "logicmacro.h"
+
+#endif
+
+
+
